pyproject.toml
[project]
name = "ollama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.12",
    "fastapi>=0.115.9",
    "llama-cloud-services>=0.6.31",
    "llama-index>=0.12.41",
    "llama-index-core>=0.12.41",
    "llama-index-embeddings-huggingface>=0.5.4",
    "llama-index-readers-json>=0.3.0",
    "llama-index-vector-stores-chroma>=0.4.2",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]
